web/requirements.txt
django>=3.2,<4.0
gunicorn
requests
psycopg2-binary
django-dotenv==1.3.0
django-storages
boto3
django-cors-headers
